Refactor remote address handler
